moved all user types to one file
